Integer infix parameter of Sharp-BQ reader macro is made optional.
Sharp-Backquote reader macro from Let Over Lambda requires us to specify
an integer infix parameter that denotes how many anaphoras should be
introduced in the resulting lambda expression. I modified this behavior
to make the parameter optional.

The main difficulty was the treatment of comma. It is treated as a
special strucutre SB-IMPL::COMMA in SBCL, while it is nothing special
object in Clozure CL. I have not yet checked about the other
implementations. In SBCL therefore, we have to strip the COMMA structure
to properly collect the anaphoras.

More accurately speaking, the reader macro behaves in a different manner
if the optional integer parameter is given. If the parameter is not
given, then the part that follows to #' is READ and the anaphoras are
extracted. Anaphoras that appears in the expression is actually declared
as the parameter of resulting lambda expression. Although anaphoras are
sorted by the suffix, it can be desultory as follows:

    #`(list ,a7 a0) => (lambda (a0 a7) `(list ,a7 ,a0))

If the optional integer parameter is given, on the other hand, it is
recognized as the number of anaphoras. Even if some of introduced
anaphoras not appear in the expression, it is introduced.

    #3`(list ,a0) => (lambda (a0 a1 a2) `(list ,a0))

The suffix starts from 0. This is another point that differes from
original version in Let Over Lambda.

The intern timing of anaphoras is different between these 2 behaviors.
READ function will interns the anaphoras in the former case while the
reader macro explicitly interns anaphoras in the latter case. The
package is not specified even in the latter case and thus it works as
expected. Even if a symbol who can be considered as an anaphora was
imported, both apperance (in parameter list and in expression) would
refers an exactly same symbol.

Any symbol whose SYMBOL-NAME is a member of following regular language
is recognized as an anaphora:

    A0 | A[1-9][0-9]*

Please be aware that the resolution is weakened since it used the
SYMBOL-NAME to detect anaphoras. In other words, the package is not
considered.

I believe that the resolution should be retrieved, but I do not have a
good approach to achive that.

I realized that PAPPLY project holds a bug in ANAPHORA-LIST during this
challenge.