A new anaphora collecting algorithm using local reader macro.
The new #` reader macro first READs one form that follows to #` with
establishing a non-terminating reader macro for #\a. The function for
character #\a READs one another form and registers a symbol if the read
object was written in decimal, or it behaves as if it was an alphadigit
character.

This approach avoids to destroy the structure and as a result it
preserves the semantics. Although the current implementation avoids to
collect the anaphoras in the nested #` reader macro, the implmenation
which collects them is also easy. The important is, the semantics.

There are a few known-problems. First, The local reader macro is only
set to #\a, but naturally #\A should also be set the same reader macro
function in typical use case. It should be decided based on the value of
(readtable-case *readtable*).

Second, the check about the suffix part of character #\a is too simple
and poor. As it was stated in the deleted part, anaphora should be the
member of follwing regular language:

    A[1-9][0-9]* | A0

Padding 0 should not be allowed because it is not clear which should be
former if both ``A0'' and ``A00'' weer used simultaneously. Moreover,
current check accepts ``A1D"" as an anaphora.

Third, this reader macro function makes a copy of current readtable.
Typically, its some sort of structure that has slots for all 128 base
characters and properties such as readtable-case. The cost for copy is
not small enough to ignore.