[Bugfix] SORT sorted the anaphoras in wrong order because of STRING<=.
A20 is considered to be ``smaller'' than A9 from the STRING<= view
point.

Since the prefix for anaphoras is the character ``A'', we can parse the
symbol-name of anaphoras by specifying :RADIX 11 to PARSE-INTEGER
instead of using SUBSEQ. I roughly checked the performance and could not
found a big difference.