WITH-ANAPHORA-PICKING to establish an anaphora picking environment.
Both of #\A and #\a must be set the same reader macro function. This
macro binds *READTABLE* to a copy of current readtable and set the
anaphora picking function to all the given characters. The picked
anaphoras are collected to the specified variable.

Another point of consern is the sideeffect to the locality of reader
macros. Since the introduced WITH-ANAPHORA-PICKING affects to a copy of
readtable, the modification made to the readtable during the read by #`
reader macro will be reverted when the read finishes. It means that any
modification of the readtable made by other reader macros cannot have a
global effect. Therefore, a reader macro intended to modify the
readtable globally will never work properly in #` forms.