Outermost #` introduces all the shared anaphras, A!0 etc.
This is an implementation of an idea to offer a way to share anaphoras
among nested abackquotes. Normal anaphoras are introduced by innermost
abackquote for those anaphoras and available only in the abackquote.
Contrary, an ``shared'' anaphora is introduced by the outermost
abackquote even if there are some abackquotes which wraps that anaphora.

In this implementation, shared anaphora is specified by the naming
convension. A symbol which is the member of following regular language
is considered to be an ``shared'' anaphora.

    A!0 | A![1-9][0-9]

And a normal anaphora is a member of following language.

    A0 | A[1-9][0-9]

(Padding zeros are available for now, but it is not intended. Bug.)

``Shared'' anaphoras are given names without ``!''. An ``shared''
anaphora A0 for instance, will be written as A0 or A!0 in the outermost
abackquote, and A!0 in the nested abackquotes. The mixed use of A!0 and
A0 result into a complicated result. They means same anaphora A0 in the
case of outermost abackquote. But in the nested abackquotes case, both
of the innermost abackquote and outermost abackquote introduces anaphora
A0, and the appearance of A!0 will result into A0. The A0 introduced by
the outermost abackquote is not visible from the inside of innermost
abackquote, and therefore, unused anaphora could be introduced. Here is
the example.

    #`(list #`(foo ,a0 ,a!0))
    ; => (LAMBDA (A0) (LAMBDA (A0) (FOO A0 A0)))

In order to avoid this situation, the user should rename the inner A0 to
different name such as A1or A2, etc. Abackquote never introduces the
inapparent anaphoras if the number argument for reader macro is omitted.

    #`(list #`(foo ,a1 ,a!0))
    ;=> (LAMBDA (A0) (LAMBDA (A1) (FOO A1 A0)))

or,

    #`(list #`(foo ,a0 ,a!1))
    ;=> (LAMBDA (A1) (LAMBDA (A0) (FOO A0 A1)))