Nested shared anaphora is impelmented.
Nested shared anaphora is an anaphora that we can control which
abackquote introduces that anaphora. In the implementation, the number
of ! controls the place of anaphora introduction. Each anaphora is
introduced by the ! numbers outer abackquote than the innermost
abackquote for the anaphora. Example:

    #`(list ,a1 ,a0 #`(foo ,a0 ,a!2 #`(bar ,a!0 ,a!!2)))
    ; => (LAMBDA (A0 A1 A2)
           `(LIST ,A1 ,A0 (LAMBDA (A0)
                            `(FOO ,A0 ,A2 (LAMBDA () `(BAR ,A0 ,A2))))))

As it was so in the previous idea, the name of anaphoras do not include
!. This is reasonable in this approach because the description of same
nested shared anaphoras can differ if the location is different. A2 of
above example is the case.